refactor(tooltip): extract shared renderFrontmatter utility and auto-linkify any http URL
- Extract duplicate renderFrontmatter into src/renderer/utils/frontmatter-renderer.ts
- Both FrontmatterPopupManager and PopupManager now share the same implementation
- Linkify any tooltip value starting with 'http', not just 'reference:' key values
