+++ src/renderer/frontmatter-popup-manager.ts
@@ -4,6 +4,7 @@
  */
 
 import { calculatePopupPosition, applyPopupPosition } from './utils/popup-position-calculator';
+import { renderFrontmatter } from './utils/frontmatter-renderer';
 import { UI_TIMING } from '../constants';
 
 interface AgentSkillItemLike {
@@ -146,60 +147,6 @@ export class FrontmatterPopupManager {
     }
   }
 
-  /**
-   * Render frontmatter content with clickable reference links
-   */
-  private renderFrontmatter(container: HTMLElement, frontmatter: string): void {
-    const lines = frontmatter.split('\n');
-
-    for (const line of lines) {
-      const colonIndex = line.indexOf(':');
-      if (colonIndex === -1) {
-        // Plain text line
-        const textNode = document.createTextNode(line);
-        container.appendChild(textNode);
-        container.appendChild(document.createElement('br'));
-        continue;
-      }
-
-      const key = line.substring(0, colonIndex).trim();
-      const value = line.substring(colonIndex + 1).trim();
-
-      // Create line container
-      const lineDiv = document.createElement('div');
-      lineDiv.className = 'frontmatter-line';
-
-      // Add key
-      const keySpan = document.createElement('span');
-      keySpan.className = 'frontmatter-key';
-      keySpan.textContent = key + ': ';
-      lineDiv.appendChild(keySpan);
-
-      // Check if value is a URL (for reference field)
-      if (key === 'reference' && value.startsWith('http')) {
-        const link = document.createElement('a');
-        link.href = value;
-        link.textContent = value;
-        link.className = 'frontmatter-link';
-        link.target = '_blank';
-        link.rel = 'noopener noreferrer';
-        // Handle click to open in external browser
-        link.addEventListener('click', (e) => {
-          e.preventDefault();
-          e.stopPropagation();
-          window.electronAPI?.shell?.openExternal?.(value);
-        });
-        lineDiv.appendChild(link);
-      } else {
-        const valueSpan = document.createElement('span');
-        valueSpan.className = 'frontmatter-value';
-        valueSpan.textContent = value;
-        lineDiv.appendChild(valueSpan);
-      }
-
-      container.appendChild(lineDiv);
-    }
-  }
 
   /**
    * Show the frontmatter popup for a command
@@ -219,7 +166,7 @@ export class FrontmatterPopupManager {
     // Create content container with parsed frontmatter
     const contentDiv = document.createElement('div');
     contentDiv.className = 'frontmatter-content';
-    this.renderFrontmatter(contentDiv, command.tooltip);
+    renderFrontmatter(contentDiv, command.tooltip);
 
     // Add file path line as last frontmatter item (before hint)
     this.addFilePathLine(contentDiv, command);
