+++ src/renderer/mentions/managers/popup-manager.ts
@@ -10,6 +10,7 @@
 
 import type { AgentItem } from '../../../types';
 import { calculatePopupPosition, applyPopupPosition } from '../../utils/popup-position-calculator';
+import { renderFrontmatter } from '../../utils/frontmatter-renderer';
 import { UI_TIMING } from '../../../constants';
 
 /**
@@ -165,55 +166,6 @@ export class PopupManager {
     }
   }
 
-  /**
-   * Render frontmatter content with clickable reference links
-   */
-  private renderFrontmatter(container: HTMLElement, frontmatter: string): void {
-    const lines = frontmatter.split('\n');
-
-    for (const line of lines) {
-      const colonIndex = line.indexOf(':');
-      if (colonIndex === -1) {
-        const textNode = document.createTextNode(line);
-        container.appendChild(textNode);
-        container.appendChild(document.createElement('br'));
-        continue;
-      }
-
-      const key = line.substring(0, colonIndex).trim();
-      const value = line.substring(colonIndex + 1).trim();
-
-      const lineDiv = document.createElement('div');
-      lineDiv.className = 'frontmatter-line';
-
-      const keySpan = document.createElement('span');
-      keySpan.className = 'frontmatter-key';
-      keySpan.textContent = key + ': ';
-      lineDiv.appendChild(keySpan);
-
-      if (key === 'reference' && value.startsWith('http')) {
-        const link = document.createElement('a');
-        link.href = value;
-        link.textContent = value;
-        link.className = 'frontmatter-link';
-        link.target = '_blank';
-        link.rel = 'noopener noreferrer';
-        link.addEventListener('click', (e) => {
-          e.preventDefault();
-          e.stopPropagation();
-          window.electronAPI?.shell?.openExternal?.(value);
-        });
-        lineDiv.appendChild(link);
-      } else {
-        const valueSpan = document.createElement('span');
-        valueSpan.className = 'frontmatter-value';
-        valueSpan.textContent = value;
-        lineDiv.appendChild(valueSpan);
-      }
-
-      container.appendChild(lineDiv);
-    }
-  }
 
   /**
    * Show frontmatter popup for an agent
@@ -234,7 +186,7 @@ export class PopupManager {
     // Create content container with parsed frontmatter (clickable reference links)
     const contentDiv = document.createElement('div');
     contentDiv.className = 'frontmatter-content';
-    this.renderFrontmatter(contentDiv, agent.tooltip);
+    renderFrontmatter(contentDiv, agent.tooltip);
 
     // Add file path line after frontmatter content
     this.addFilePathLine(contentDiv, agent.filePath);
